Remove `BasePlatformOption.Platform` default
When `sqlserver` is provided as the default value for the command-line option, the target platform is already set when checked on `CommandLineService.cs` line 92.
This prevents the environment variable YUNIQL_TARGET_PLATFORM from being able to provide a different default value on `CommandLineService.cs` line 94.
By removing the default value from the CLI option definition, the environment variable lookup in `CommandLineService.cs` works as expected.